feat: event-based buffered checks
