+++ demo/index.html
@@ -52,7 +52,7 @@
             </audio>
 
             <audio controls preload="auto">
-                <source src="./audio-examples/long-audio-example.mp3" />
+                <source src="./audio-examples/long-audio-example.mp3" preload="true"/>
                 Your browser does not support the <code>audio</code> element.
             </audio>
         </div>
